Refactor headers & sources to be compatible with new project structure
