+++ src/client/connection.c
@@ -5,12 +5,12 @@
 #include "client/types/types.h"
 #include "client/types/string.h"
 
-#include "client/cerver/network.h"
-#include "client/cerver/cerver.h"
-#include "client/cerver/client.h"
-#include "client/cerver/connection.h"
-#include "client/cerver/handler.h"
-#include "client/cerver/packets.h"
+#include "client/network.h"
+#include "client/cerver.h"
+#include "client/client.h"
+#include "client/connection.h"
+#include "client/handler.h"
+#include "client/packets.h"
 
 #include "client/threads/thread.h"
